🚚 Move Template-related enums in tqec.templates
