+++ src/tqec/__init__.py
@@ -12,10 +12,7 @@ from .circuit import (
     merge_scheduled_circuits,
 )
 from .enums import (
-    CornerPositionEnum,
     PlaquetteOrientation,
-    TemplateOrientation,
-    TemplateRelativePositionEnum,
 )
 from .exceptions import TQECException
 from .noise_models import (
@@ -53,4 +50,9 @@ from .templates import (
     TemplateWithIndices,
     display_template,
     display_templates_svg,
+)
+from .templates.enums import (
+    CornerPositionEnum,
+    TemplateOrientation,
+    TemplateRelativePositionEnum,
 )
