+++ src/tqec/plaquette/qubit.py
@@ -5,8 +5,9 @@ from dataclasses import dataclass
 
 import cirq
 
-from tqec.enums import PlaquetteOrientation, PlaquetteSide, TemplateOrientation
+from tqec.enums import PlaquetteOrientation, PlaquetteSide
 from tqec.position import Position
+from tqec.templates.enums import TemplateOrientation
 
 
 @dataclass(frozen=True)
